Rename files with .h extensions to use .hpp
With the rationale that pure c++ files should use `.hpp`, and only mixed c/c++ implementation (stuff that could be compiled as c or c++ should have .h headers)
