+++ flow/include/captor/polling.hpp
@@ -2,28 +2,45 @@
  * @copyright 2020 Fetch Robotics Inc.
  * @author Brian Cairl
  */
-#ifndef FLOW_CAPTURE_CAPTOR_NOLOCK_H
-#define FLOW_CAPTURE_CAPTOR_NOLOCK_H
+#ifndef FLOW_CAPTURE_CAPTOR_POLLING_H
+#define FLOW_CAPTURE_CAPTOR_POLLING_H
 
 // C++ Standard Library
 #include <algorithm>
 #include <iterator>
-#include <memory>
 #include <mutex>
 #include <type_traits>
 
 // Flow
-#include <flow/captor.h>
+#include <flow/captor.hpp>
 
 namespace flow
 {
 
 /**
- * @copydoc Captor
- * @note No-lock captor implementation meant for polling with <code>Captor::capture</code> in single-threaded code.
+ * @brief Stand-in type used to signify that captors will be used in a threaded context, but will not wait for data
+ *
+ *        Captors use <code>BasicLockableT</code> to protect data input/output from the capture queue, but
+ *        do not wait on a condition variable for new inputs before attempting to run a synchronization policy.
+ *        This allows for polling with the <code>Captor::capture</code> method
+ * \n
+ *        See https://en.cppreference.com/w/cpp/named_req/TimedLockable for more information on
+ * <code>BasicLockableT</code> criteria
  */
-template <typename CaptorT, typename QueueMonitorT>
-class Captor<CaptorT, NoLock, QueueMonitorT> : public CaptorInterface<Captor<CaptorT, NoLock, QueueMonitorT>>
+template <typename BasicLockableT = std::lock_guard<std::mutex>> struct PollingLock
+{
+  using type = BasicLockableT;
+};
+
+
+/**
+ * @copydoc Captor
+ * @note Implements a simple locking policy using a <code>BasicLockableT</code> which
+ *       protects data input/output, meant for polling with <code>Captor::capture</code>.
+ */
+template <typename CaptorT, typename BasicLockableT, typename QueueMonitorT>
+class Captor<CaptorT, PollingLock<BasicLockableT>, QueueMonitorT>
+    : public CaptorInterface<Captor<CaptorT, PollingLock<BasicLockableT>, QueueMonitorT>>
 {
 public:
   /// Data dispatch type
@@ -62,6 +79,8 @@ private:
    */
   inline void reset_impl()
   {
+    BasicLockableT lock{queue_mutex_};
+
     // Run reset behavior specific to this captor
     derived()->reset_policy_impl();
 
@@ -74,6 +93,8 @@ private:
    */
   inline void abort_impl(const stamp_type& t_abort)
   {
+    BasicLockableT lock{queue_mutex_};
+
     // Run abort behavior specific to this captor
     derived()->abort_policy_impl(t_abort);
   }
@@ -81,13 +102,18 @@ private:
   /**
    * @copydoc CaptorInterface::size
    */
-  inline size_type size_impl() const { return CaptorInterfaceType::queue_.size(); }
+  inline size_type size_impl() const
+  {
+    BasicLockableT lock{queue_mutex_};
+    return CaptorInterfaceType::queue_.size();
+  }
 
   /**
    * @copydoc CaptorInterface::inject
    */
   template <typename... DispatchConstructorArgTs> inline void inject_impl(DispatchConstructorArgTs&&... dispatch_args)
   {
+    BasicLockableT lock{queue_mutex_};
     CaptorInterfaceType::insert_and_limit(std::forward<DispatchConstructorArgTs>(dispatch_args)...);
   }
 
@@ -97,6 +123,7 @@ private:
   template <typename FirstForwardDispatchIteratorT, typename LastForwardDispatchIteratorT>
   inline void insert_impl(FirstForwardDispatchIteratorT first, LastForwardDispatchIteratorT last)
   {
+    BasicLockableT lock{queue_mutex_};
     std::for_each(
       first, last, [this](const DispatchType& dispatch) { CaptorInterfaceType::insert_and_limit(dispatch); });
   }
@@ -106,6 +133,8 @@ private:
    */
   inline void remove_impl(const stamp_type& t_remove)
   {
+    BasicLockableT lock{queue_mutex_};
+
     // Remove all data before this time
     CaptorInterfaceType::queue_.remove_before(t_remove);
   }
@@ -113,18 +142,27 @@ private:
   /**
    * @copydoc CaptorInterface::set_capacity
    */
-  inline void set_capacity_impl(const size_type capacity) { CaptorInterfaceType::capacity_ = capacity; }
+  inline void set_capacity_impl(const size_type capacity)
+  {
+    BasicLockableT lock{queue_mutex_};
+    CaptorInterfaceType::capacity_ = capacity;
+  }
 
   /**
    * @copydoc CaptorInterface::get_capacity
    */
-  inline size_type get_capacity_impl() const { return CaptorInterfaceType::capacity_; }
+  inline size_type get_capacity_impl() const
+  {
+    BasicLockableT lock{queue_mutex_};
+    return CaptorInterfaceType::capacity_;
+  }
 
   /**
    * @copydoc CaptorInterface::get_available_stamp_range
    */
   inline CaptureRange<stamp_type> get_available_stamp_range_impl() const
   {
+    BasicLockableT lock{queue_mutex_};
     return queue_.empty() ? CaptureRange<stamp_type>{}
                           : CaptureRange<stamp_type>{queue_.oldest_stamp(), queue_.newest_stamp()};
   }
@@ -135,6 +173,7 @@ private:
   template <typename OutputDispatchIteratorT, typename CaptureRangeT>
   inline State capture_impl(OutputDispatchIteratorT&& output, CaptureRangeT&& range)
   {
+    BasicLockableT lock{queue_mutex_};
     return derived()->capture_policy_impl(
       std::forward<OutputDispatchIteratorT>(output), std::forward<CaptureRangeT>(range));
   }
@@ -142,8 +181,9 @@ private:
   /**
    * @copydoc CaptorInterface::dry_capture
    */
-  template <typename CaptureRangeT> State dry_capture_impl(CaptureRangeT&& range)
+  template <typename CaptureRangeT> inline State dry_capture_impl(CaptureRangeT&& range) const
   {
+    BasicLockableT lock{queue_mutex_};
     return derived()->dry_capture_policy_impl(std::forward<CaptureRangeT>(range));
   }
 
@@ -152,6 +192,7 @@ private:
    */
   template <typename InpectCallbackT> void inspect_impl(InpectCallbackT&& inspect_dispatch_cb) const
   {
+    BasicLockableT lock{queue_mutex_};
     for (const auto& dispatch : CaptorInterfaceType::queue_)
     {
       inspect_dispatch_cb(dispatch);
@@ -159,15 +200,19 @@ private:
   }
 
   /**
-   * @copydoc CaptorInterface::update_queue_monitor
+   * @copydoc CaptorInterface::update_monitor
    */
   template <typename CaptureRangeT> void update_queue_monitor_impl(CaptureRangeT&& range, const State sync_state)
   {
+    BasicLockableT lock{queue_mutex_};
     CaptorInterfaceType::queue_monitor_.update(
       CaptorInterfaceType::queue_, std::forward<CaptureRangeT>(range), sync_state);
   }
 
-  using CaptorInterfaceType = CaptorInterface<Captor<CaptorT, NoLock, QueueMonitorT>>;
+  /// Mutex to protect queue ONLY
+  mutable std::mutex queue_mutex_;
+
+  using CaptorInterfaceType = CaptorInterface<Captor<CaptorT, PollingLock<BasicLockableT>, QueueMonitorT>>;
   friend CaptorInterfaceType;
 
   FLOW_IMPLEMENT_CRTP_BASE(CaptorT);
@@ -176,6 +221,15 @@ protected:
   using CaptorInterfaceType::queue_;
 };
 
+
+/**
+ * @copydoc is_polling_lock
+ *
+ * @note true case
+ */
+template <typename LockableT> struct is_polling_lock<PollingLock<LockableT>> : std::integral_constant<bool, true>
+{};
+
 }  // namespace flow
 
-#endif  // FLOW_CAPTURE_CAPTOR_NOLOCK_H
+#endif  // FLOW_CAPTURE_CAPTOR_POLLING_H
