+++ flow/include/dispatch/pair.hpp
@@ -11,7 +11,7 @@
 #include <utility>
 
 // Flow
-#include <flow/dispatch.h>
+#include <flow/dispatch.hpp>
 
 namespace flow
 {
