+++ flow/include/dispatch_ostream.hpp
@@ -1,17 +1,15 @@
 /**
  * @copyright 2020 Fetch Robotics Inc.
  * @author Brian Cairl
- *
- * @file dispatch_ostream.h
  */
-#ifndef FLOW_CAPTOR_DISPATCH_OSTREAM_H
-#define FLOW_CAPTOR_DISPATCH_OSTREAM_H
+#ifndef FLOW_CAPTOR_DISPATCH_OSTREAM_HPP
+#define FLOW_CAPTOR_DISPATCH_OSTREAM_HPP
 
 // C++ Standard Library
 #include <ostream>
 
 // Flow
-#include <flow/dispatch.h>
+#include <flow/dispatch.hpp>
 
 namespace flow
 {
@@ -43,4 +41,4 @@ template <typename StampT> inline std::ostream& operator<<(std::ostream& os, con
 
 }  // namespace flow
 
-#endif  // FLOW_CAPTOR_DISPATCH_OSTREAM_H
+#endif  // FLOW_CAPTOR_DISPATCH_OSTREAM_HPP
