+++ flow/include/driver/driver.hpp
@@ -1,84 +1,81 @@
 /**
  * @copyright 2020 Fetch Robotics Inc.
  * @author Brian Cairl
- *
- * @file follower.h
  */
-#ifndef FLOW_FOLLOWER_FOLLOWER_H
-#define FLOW_FOLLOWER_FOLLOWER_H
+#ifndef FLOW_DRIVER_DRIVER_HPP
+#define FLOW_DRIVER_DRIVER_HPP
 
 // C++ Standard Library
 #include <type_traits>
 
 // Flow
-#include <flow/captor.h>
-#include <flow/captor_state.h>
+#include <flow/captor.hpp>
+#include <flow/captor_state.hpp>
 #include <flow/utility/implement_crtp_base.hpp>
 
 namespace flow
 {
 
 // Forward declaration
-template <typename PolicyT> class Follower;
+template <typename PolicyT> class Driver;
 
 
 /**
  * @copydoc CaptorTraits
  * @tparam PolicyT  CRTP-derived captor with specialized capture policy
  */
-template <typename PolicyT> struct CaptorTraits<Follower<PolicyT>> : CaptorTraits<PolicyT>
+template <typename PolicyT> struct CaptorTraits<Driver<PolicyT>> : CaptorTraits<PolicyT>
 {};
 
 
 /**
- * @brief CRTP-base for Follower input-capture policies
+ * @brief CRTP-base for Driver input-capture policies
  *
- *        Captures data w.r.t to a driving sequencing range, produced by a Driver,
- *        according to a synchronization policy
+ *        Captures data produces a synchronization sequencing range used
+ *        to synchronize data produced by Follower buffers
  *
  * @tparam PolicyT  CRTP-derived captor with specialized capture policy
  */
 template <typename PolicyT>
-class Follower : public Captor<
-                   Follower<PolicyT>,
-                   typename CaptorTraits<PolicyT>::LockPolicyType,
-                   typename CaptorTraits<PolicyT>::DispatchQueueMonitorType>
+class Driver
+    : public Captor<Driver<PolicyT>, typename CaptorTraits<PolicyT>::LockPolicyType, DefaultDispatchQueueMonitor>
 {
 public:
   /// Underlying dispatch container type
   using DispatchContainerType = typename CaptorTraits<PolicyT>::DispatchContainerType;
 
-  /// Queue monitor/capture preconditioning type
+  /// Queue monitor type
   using DispatchQueueMonitorType = typename CaptorTraits<PolicyT>::DispatchQueueMonitorType;
 
   /// Data stamp type
   using stamp_type = typename CaptorTraits<PolicyT>::stamp_type;
 
   /**
-   * @brief Initialization constructor
+   * @brief Dispatch container constructor
    *
    * @param container  container object with some initial state
    * @param queue_monitor  queue monitor with some initial state
    */
-  Follower(const DispatchContainerType& container, const DispatchQueueMonitorType& queue_monitor);
+  explicit Driver(const DispatchContainerType& container, const DispatchQueueMonitorType& queue_monitor);
 
 private:
   /**
-   * @brief Checks if buffer is in ready state and collects data based on a target time
+   * @brief Checks if buffer is in ready state and captures data
    *
    * @param[out] output  output data iterator
-   * @param[in] range  data capture/sequencing range
+   * @param[in,out] range  data capture/sequencing range
    *
    * @retval State::PRIMED    Data have been captured
    * @retval State::RETRY  Captor should continue waiting for messages after prime attempt
    */
   template <typename OutputDispatchIteratorT>
-  inline State capture_policy_impl(OutputDispatchIteratorT&& output, const CaptureRange<stamp_type>& range);
+  inline State capture_policy_impl(OutputDispatchIteratorT&& output, CaptureRange<stamp_type>& range);
 
   /**
    * @copydoc CaptorInterface::dry_capture
+   * @note Sets data capture/sequencing range for following captors
    */
-  inline State dry_capture_policy_impl(const CaptureRange<stamp_type>& range);
+  inline State dry_capture_policy_impl(CaptureRange<stamp_type>& range);
 
   /**
    * @brief Defines Captor behavior on <code>ABORT</code>
@@ -96,29 +93,25 @@ private:
 
   FLOW_IMPLEMENT_CRTP_BASE(PolicyT);
 
-  using CaptorType = Captor<
-    Follower,
-    typename CaptorTraits<PolicyT>::LockPolicyType,
-    typename CaptorTraits<PolicyT>::DispatchQueueMonitorType>;
+  using CaptorType = Captor<Driver, typename CaptorTraits<PolicyT>::LockPolicyType, DefaultDispatchQueueMonitor>;
   friend CaptorType;
 
 protected:
   using CaptorType::queue_;
-  using CaptorType::queue_monitor_;
 };
 
 
 /**
- * @brief Checks if captor object derived from a Follower base
+ * @brief Checks if captor object derived from a Driver base
  * @param CaptorT  object to test
  */
 template <typename CaptorT>
-struct is_follower : std::integral_constant<bool, std::is_base_of<Follower<CaptorT>, CaptorT>::value>
+struct is_driver : std::integral_constant<bool, std::is_base_of<Driver<CaptorT>, CaptorT>::value>
 {};
 
 }  // namespace flow
 
 // Flow (implementation)
-#include "flow/src/follower/follower.hpp"
+#include "flow/src/driver/driver.hpp"
 
-#endif  // FLOW_FOLLOWER_FOLLOWER_H
+#endif  // FLOW_DRIVER_DRIVER_HPP
