--- flow/include/drivers.h
@@ -1,26 +1,0 @@
-/**
- * @copyright 2020 Fetch Robotics Inc.
- * @author Brian Cairl
- *
- * @file drivers.h
- */
-#ifndef FLOW_DRIVERS_H
-#define FLOW_DRIVERS_H
-
-// Flow
-#include <flow/driver/batch.h>
-#include <flow/driver/chunk.h>
-#include <flow/driver/next.h>
-#include <flow/driver/throttled.h>
-
-namespace flow
-{
-/**
- * @brief Synchronizations buffer with policies which produce sequencing ranges
- */
-namespace driver
-{}
-
-}  // flow
-
-#endif  // FLOW_DRIVERS_H
