--- flow/include/flow.h
@@ -1,21 +1,0 @@
-/**
- * @copyright 2020 Fetch Robotics Inc.
- * @author Brian Cairl
- *
- * @file flow.h
- */
-#ifndef FLOW_FLOW_H
-#define FLOW_FLOW_H
-
-// Flow
-#include <flow/drivers.h>
-#include <flow/followers.h>
-#include <flow/synchronizer.h>
-
-/**
- * @brief Flow input synchronization library components
- */
-namespace flow
-{}
-
-#endif  // FLOW_FLOW_H
