+++ flow/include/flow.hpp
@@ -1,0 +1,19 @@
+/**
+ * @copyright 2020 Fetch Robotics Inc.
+ * @author Brian Cairl
+ */
+#ifndef FLOW_FLOW_HPP
+#define FLOW_FLOW_HPP
+
+// Flow
+#include <flow/drivers.hpp>
+#include <flow/followers.hpp>
+#include <flow/synchronizer.hpp>
+
+/**
+ * @brief Flow input synchronization library components
+ */
+namespace flow
+{}
+
+#endif  // FLOW_FLOW_HPP
