+++ flow/include/follower/any_before.hpp
@@ -1,14 +1,12 @@
 /**
  * @copyright 2020 Fetch Robotics Inc.
- * @author Brian Cairl, Derek King
- *
- * @file closest_before.h
+ * @author Levon Avagyan, Brian Cairl
  */
-#ifndef FLOW_FOLLOWER_CLOSEST_BEFORE_H
-#define FLOW_FOLLOWER_CLOSEST_BEFORE_H
+#ifndef FLOW_FOLLOWER_ANY_BEFORE_HPP
+#define FLOW_FOLLOWER_ANY_BEFORE_HPP
 
 // Flow
-#include <flow/follower/follower.h>
+#include <flow/follower/follower.hpp>
 
 namespace flow
 {
@@ -16,9 +14,18 @@ namespace follower
 {
 
 /**
- * @brief Captures one element before the capture range lower bound, minus a delay period, within an expected period.
+ * @brief Captures all data elements from a delay before the driving sequencing stamp
  *
- *        All older elements are removed.
+ *        This capture buffer will capture data which is behind the driving upper
+ *        sequence stamp (<code>range.upper_stamp</code>) by some sequencing delay
+ *        w.r.t a driver-provided target time. It will return all data at and before
+ *        that sequencing boundary that has not previously been captured.
+ * \n
+ *        This capture buffer is always ready, and will always return with a PRIMED state,
+ *        regardless of whether or not there is data available to capture.
+ * \n
+ *        <b>Data removal:</b> Captor will remove all data before the driving time
+ *        message minus the delay
  *
  * @tparam DispatchT  data dispatch type
  * @tparam LockPolicyT  a BasicLockable (https://en.cppreference.com/w/cpp/named_req/BasicLockable) object or NoLock or
@@ -26,41 +33,38 @@ namespace follower
  * @tparam ContainerT  underlying <code>DispatchT</code> container type
  * @tparam QueueMonitorT  object used to monitor queue state on each insertion; used to precondition capture
  *
- * @warn ClosestBefore will behave non-deterministically if actual input period (difference between successive
- *       dispatch stamps) does not match the <code>period</code> argument specified on construction. For example,
- *       if <code>period</code> is too large, than multiple inputs could appear before the driving range, causing
- *       for different data on two or more iterations where the "latest" data was assumed to have been the same
+ * @warn This captor WILL NOT behave deterministically if all data is not available before capture time minus
+ *       the specified delay. As such, setting the delay properly will alleviate non-deterministic behavior.
+ *       This is the only <i>optional</i> captor, and should be used with great caution.
  */
 template <
   typename DispatchT,
   typename LockPolicyT = NoLock,
   typename ContainerT = DefaultContainer<DispatchT>,
   typename QueueMonitorT = DefaultDispatchQueueMonitor>
-class ClosestBefore : public Follower<ClosestBefore<DispatchT, LockPolicyT, ContainerT, QueueMonitorT>>
+class AnyBefore : public Follower<AnyBefore<DispatchT, LockPolicyT, ContainerT, QueueMonitorT>>
 {
 public:
   /// Data stamp type
-  using stamp_type = typename CaptorTraits<ClosestBefore>::stamp_type;
+  using stamp_type = typename CaptorTraits<AnyBefore>::stamp_type;
 
   /// Data stamp duration type
-  using offset_type = typename CaptorTraits<ClosestBefore>::offset_type;
+  using offset_type = typename CaptorTraits<AnyBefore>::offset_type;
 
   /**
    * @brief Setup constructor
    *
-   * @param period  expected half-period between successive data elements
    * @param delay  the delay with which to capture
    * @param container  container object with some initial state
    * @param queue_monitor  queue monitor with some initial state
    */
-  ClosestBefore(
-    const offset_type& period,
+  explicit AnyBefore(
     const offset_type& delay,
     const ContainerT& container = ContainerT{},
     const QueueMonitorT& queue_monitor = QueueMonitorT{});
 
 private:
-  using PolicyType = Follower<ClosestBefore<DispatchT, LockPolicyT, ContainerT, QueueMonitorT>>;
+  using PolicyType = Follower<AnyBefore<DispatchT, LockPolicyT, ContainerT, QueueMonitorT>>;
   friend PolicyType;
 
   /**
@@ -69,11 +73,7 @@ private:
    * @param[out] output  output data iterator
    * @param[in] range  data capture/sequencing range
    *
-   * @retval ABORT   If next closest element has a sequencing stamp greater than <code>range.upper_stamp</code>
-   * @retval PRIMED  If next closest element to is available
-   * @retval RETRY   Element with sequence stamp greater than <code>range.upper_stamp</code> exists, and
-   *                 there is a data element within the expected duration window before
-   *                 <code>range.upper_stamp</code>
+   * @retval PRIMED  always
    */
   template <typename OutputDispatchIteratorT>
   inline State capture_follower_impl(OutputDispatchIteratorT output, const CaptureRange<stamp_type>& range);
@@ -81,7 +81,7 @@ private:
   /**
    * @copydoc Follower::dry_capture_policy_impl
    */
-  inline State dry_capture_follower_impl(const CaptureRange<stamp_type>& range);
+  inline State dry_capture_follower_impl(const CaptureRange<stamp_type>& range) const;
 
   /**
    * @copydoc Follower::abort_policy_impl
@@ -92,9 +92,6 @@ private:
    * @copydoc Follower::reset_policy_impl
    */
   inline void reset_follower_impl() noexcept(true) {}
-
-  /// Expected update period
-  offset_type period_;
 
   /// Capture delay
   offset_type delay_;
@@ -113,7 +110,7 @@ private:
  * @tparam QueueMonitorT queue monitor/capture preconditioning type
  */
 template <typename DispatchT, typename LockPolicyT, typename ContainerT, typename QueueMonitorT>
-struct CaptorTraits<follower::ClosestBefore<DispatchT, LockPolicyT, ContainerT, QueueMonitorT>>
+struct CaptorTraits<follower::AnyBefore<DispatchT, LockPolicyT, ContainerT, QueueMonitorT>>
     : CaptorTraitsFromDispatch<DispatchT>
 {
   /// Underlying dispatch container type
@@ -125,14 +122,14 @@ struct CaptorTraits<follower::ClosestBefore<DispatchT, LockPolicyT, ContainerT, 
   /// Thread locking policy type
   using LockPolicyType = LockPolicyT;
 
-  /// Indicates that data from this captor will always be captured deterministically, so long as data
-  /// injection is monotonically sequenced
-  static constexpr bool is_capture_deterministic = true;
+  /// Indicates that data from this captor will NOT always be captured deterministically;
+  /// i.e. is always dependent on when data is injected, and when captrue is executed
+  static constexpr bool is_capture_deterministic = false;
 };
 
 }  // namespace flow
 
 // Flow (implementation)
-#include "flow/src/follower/closest_before.hpp"
+#include "flow/src/follower/any_before.hpp"
 
-#endif  // FLOW_FOLLOWER_CLOSEST_BEFORE_H
+#endif  // FLOW_FOLLOWER_ANY_BEFORE_HPP
