+++ flow/include/follower/before.hpp
@@ -1,14 +1,12 @@
 /**
  * @copyright 2020 Fetch Robotics Inc.
  * @author Levon Avagyan, Brian Cairl
- *
- * @file before.h
  */
-#ifndef FLOW_FOLLOWER_ANY_BEFORE_H
-#define FLOW_FOLLOWER_ANY_BEFORE_H
+#ifndef FLOW_FOLLOWER_BEFORE_HPP
+#define FLOW_FOLLOWER_BEFORE_HPP
 
 // Flow
-#include <flow/follower/follower.h>
+#include <flow/follower/follower.hpp>
 
 namespace flow
 {
@@ -16,42 +14,30 @@ namespace follower
 {
 
 /**
- * @brief Captures all data elements from a delay before the driving sequencing stamp
+ * @brief Captures all  elements before the capture range lower bound, minus a delay period.
  *
- *        This capture buffer will capture data which is behind the driving upper
- *        sequence stamp (<code>range.upper_stamp</code>) by some sequencing delay
- *        w.r.t a driver-provided target time. It will return all data at and before
- *        that sequencing boundary that has not previously been captured.
- * \n
- *        This capture buffer is always ready, and will always return with a PRIMED state,
- *        regardless of whether or not there is data available to capture.
- * \n
- *        <b>Data removal:</b> Captor will remove all data before the driving time
- *        message minus the delay
+ *        Once at least a single element is available after said sequencing boundary.
+ *        All of the captured elements are removed.
  *
  * @tparam DispatchT  data dispatch type
  * @tparam LockPolicyT  a BasicLockable (https://en.cppreference.com/w/cpp/named_req/BasicLockable) object or NoLock or
  * PollingLock
  * @tparam ContainerT  underlying <code>DispatchT</code> container type
  * @tparam QueueMonitorT  object used to monitor queue state on each insertion; used to precondition capture
- *
- * @warn This captor WILL NOT behave deterministically if all data is not available before capture time minus
- *       the specified delay. As such, setting the delay properly will alleviate non-deterministic behavior.
- *       This is the only <i>optional</i> captor, and should be used with great caution.
  */
 template <
   typename DispatchT,
   typename LockPolicyT = NoLock,
   typename ContainerT = DefaultContainer<DispatchT>,
   typename QueueMonitorT = DefaultDispatchQueueMonitor>
-class AnyBefore : public Follower<AnyBefore<DispatchT, LockPolicyT, ContainerT, QueueMonitorT>>
+class Before : public Follower<Before<DispatchT, LockPolicyT, ContainerT, QueueMonitorT>>
 {
 public:
   /// Data stamp type
-  using stamp_type = typename CaptorTraits<AnyBefore>::stamp_type;
+  using stamp_type = typename CaptorTraits<Before>::stamp_type;
 
   /// Data stamp duration type
-  using offset_type = typename CaptorTraits<AnyBefore>::offset_type;
+  using offset_type = typename CaptorTraits<Before>::offset_type;
 
   /**
    * @brief Setup constructor
@@ -60,13 +46,13 @@ public:
    * @param container  container object with some initial state
    * @param queue_monitor  queue monitor with some initial state
    */
-  explicit AnyBefore(
+  explicit Before(
     const offset_type& delay,
     const ContainerT& container = ContainerT{},
     const QueueMonitorT& queue_monitor = QueueMonitorT{});
 
 private:
-  using PolicyType = Follower<AnyBefore<DispatchT, LockPolicyT, ContainerT, QueueMonitorT>>;
+  using PolicyType = Follower<Before<DispatchT, LockPolicyT, ContainerT, QueueMonitorT>>;
   friend PolicyType;
 
   /**
@@ -75,7 +61,9 @@ private:
    * @param[out] output  output data iterator
    * @param[in] range  data capture/sequencing range
    *
-   * @retval PRIMED  always
+   * @retval PRIMED  if there is a Dispatch element with a sequencing stamp greater than or
+   *                 equal to the upper driving stamp, minus specified delay
+   * @retval RETRY  otherwise
    */
   template <typename OutputDispatchIteratorT>
   inline State capture_follower_impl(OutputDispatchIteratorT output, const CaptureRange<stamp_type>& range);
@@ -112,7 +100,7 @@ private:
  * @tparam QueueMonitorT queue monitor/capture preconditioning type
  */
 template <typename DispatchT, typename LockPolicyT, typename ContainerT, typename QueueMonitorT>
-struct CaptorTraits<follower::AnyBefore<DispatchT, LockPolicyT, ContainerT, QueueMonitorT>>
+struct CaptorTraits<follower::Before<DispatchT, LockPolicyT, ContainerT, QueueMonitorT>>
     : CaptorTraitsFromDispatch<DispatchT>
 {
   /// Underlying dispatch container type
@@ -124,14 +112,14 @@ struct CaptorTraits<follower::AnyBefore<DispatchT, LockPolicyT, ContainerT, Queu
   /// Thread locking policy type
   using LockPolicyType = LockPolicyT;
 
-  /// Indicates that data from this captor will NOT always be captured deterministically;
-  /// i.e. is always dependent on when data is injected, and when captrue is executed
-  static constexpr bool is_capture_deterministic = false;
+  /// Indicates that data from this captor will always be captured deterministically, so long as data
+  /// injection is monotonically sequenced
+  static constexpr bool is_capture_deterministic = true;
 };
 
 }  // namespace flow
 
 // Flow (implementation)
-#include "flow/src/follower/any_before.hpp"
+#include "flow/src/follower/before.hpp"
 
-#endif  // FLOW_FOLLOWER_ANY_BEFORE_H
+#endif  // FLOW_FOLLOWER_BEFORE_HPP
