+++ flow/include/follower/count_before.hpp
@@ -1,14 +1,12 @@
 /**
  * @copyright 2020 Fetch Robotics Inc.
- * @author Levon Avagyan, Brian Cairl
- *
- * @file before.h
+ * @author Brian Cairl
  */
-#ifndef FLOW_FOLLOWER_BEFORE_H
-#define FLOW_FOLLOWER_BEFORE_H
+#ifndef FLOW_FOLLOWER_COUNT_BEFORE_HPP
+#define FLOW_FOLLOWER_COUNT_BEFORE_HPP
 
 // Flow
-#include <flow/follower/follower.h>
+#include <flow/follower/follower.hpp>
 
 namespace flow
 {
@@ -16,10 +14,9 @@ namespace follower
 {
 
 /**
- * @brief Captures all  elements before the capture range lower bound, minus a delay period.
+ * @brief Captures N-elements before the capture range lower bound, minus a delay period
  *
- *        Once at least a single element is available after said sequencing boundary.
- *        All of the captured elements are removed.
+ *        All older elements are removed.
  *
  * @tparam DispatchT  data dispatch type
  * @tparam LockPolicyT  a BasicLockable (https://en.cppreference.com/w/cpp/named_req/BasicLockable) object or NoLock or
@@ -32,29 +29,36 @@ template <
   typename LockPolicyT = NoLock,
   typename ContainerT = DefaultContainer<DispatchT>,
   typename QueueMonitorT = DefaultDispatchQueueMonitor>
-class Before : public Follower<Before<DispatchT, LockPolicyT, ContainerT, QueueMonitorT>>
+class CountBefore : public Follower<CountBefore<DispatchT, LockPolicyT, ContainerT, QueueMonitorT>>
 {
 public:
+  /// Integer size type
+  using size_type = typename CaptorTraits<CountBefore>::size_type;
+
   /// Data stamp type
-  using stamp_type = typename CaptorTraits<Before>::stamp_type;
+  using stamp_type = typename CaptorTraits<CountBefore>::stamp_type;
 
   /// Data stamp duration type
-  using offset_type = typename CaptorTraits<Before>::offset_type;
+  using offset_type = typename CaptorTraits<CountBefore>::offset_type;
 
   /**
    * @brief Setup constructor
    *
+   * @param count  number of elements before to capture
    * @param delay  the delay with which to capture
    * @param container  container object with some initial state
    * @param queue_monitor  queue monitor with some initial state
+   *
+   * @throws <code>std::invalid_argument</code> if <code>count == 0</code>
    */
-  explicit Before(
+  CountBefore(
+    const size_type count,
     const offset_type& delay,
     const ContainerT& container = ContainerT{},
     const QueueMonitorT& queue_monitor = QueueMonitorT{});
 
 private:
-  using PolicyType = Follower<Before<DispatchT, LockPolicyT, ContainerT, QueueMonitorT>>;
+  using PolicyType = Follower<CountBefore<DispatchT, LockPolicyT, ContainerT, QueueMonitorT>>;
   friend PolicyType;
 
   /**
@@ -63,8 +67,9 @@ private:
    * @param[out] output  output data iterator
    * @param[in] range  data capture/sequencing range
    *
-   * @retval PRIMED  if there is a Dispatch element with a sequencing stamp greater than or
+   * @retval PRIMED  if there are N Dispatch elements with a sequencing stamp greater than or
    *                 equal to the upper driving stamp, minus specified delay
+   * @retval ABORT  if capture is not possible
    * @retval RETRY  otherwise
    */
   template <typename OutputDispatchIteratorT>
@@ -73,17 +78,20 @@ private:
   /**
    * @copydoc Follower::dry_capture_policy_impl
    */
-  inline State dry_capture_follower_impl(const CaptureRange<stamp_type>& range) const;
+  inline State dry_capture_follower_impl(const CaptureRange<stamp_type>& range);
 
   /**
    * @copydoc Follower::abort_policy_impl
    */
-  inline void abort_follower_impl(const stamp_type& t_abort);
+  constexpr void abort_follower_impl(const stamp_type& t_abort) noexcept(true) {}
 
   /**
    * @copydoc Follower::reset_policy_impl
    */
-  inline void reset_follower_impl() noexcept(true) {}
+  constexpr void reset_follower_impl() noexcept(true) {}
+
+  /// Number of elements to be caputed
+  size_type count_;
 
   /// Capture delay
   offset_type delay_;
@@ -102,7 +110,7 @@ private:
  * @tparam QueueMonitorT queue monitor/capture preconditioning type
  */
 template <typename DispatchT, typename LockPolicyT, typename ContainerT, typename QueueMonitorT>
-struct CaptorTraits<follower::Before<DispatchT, LockPolicyT, ContainerT, QueueMonitorT>>
+struct CaptorTraits<follower::CountBefore<DispatchT, LockPolicyT, ContainerT, QueueMonitorT>>
     : CaptorTraitsFromDispatch<DispatchT>
 {
   /// Underlying dispatch container type
@@ -114,14 +122,14 @@ struct CaptorTraits<follower::Before<DispatchT, LockPolicyT, ContainerT, QueueMo
   /// Thread locking policy type
   using LockPolicyType = LockPolicyT;
 
-  /// Indicates that data from this captor will always be captured deterministically, so long as data
-  /// injection is monotonically sequenced
-  static constexpr bool is_capture_deterministic = true;
+  /// Indicates that data from this captor will NOT always be captured deterministically;
+  /// i.e. is always dependent on when data is injected, and when captrue is executed
+  static constexpr bool is_capture_deterministic = false;
 };
 
 }  // namespace flow
 
 // Flow (implementation)
-#include "flow/src/follower/before.hpp"
+#include "flow/src/follower/count_before.hpp"
 
-#endif  // FLOW_FOLLOWER_BEFORE_H
+#endif  // FLOW_FOLLOWER_COUNT_BEFORE_HPP
