+++ flow/include/follower/ranged.hpp
@@ -1,14 +1,12 @@
 /**
  * @copyright 2020 Fetch Robotics Inc.
  * @author Brian Cairl
- *
- * @file count_before.h
  */
-#ifndef FLOW_FOLLOWER_COUNT_BEFORE_H
-#define FLOW_FOLLOWER_COUNT_BEFORE_H
+#ifndef FLOW_FOLLOWER_RANGED_HPP
+#define FLOW_FOLLOWER_RANGED_HPP
 
 // Flow
-#include <flow/follower/follower.h>
+#include <flow/follower/follower.hpp>
 
 namespace flow
 {
@@ -16,9 +14,10 @@ namespace follower
 {
 
 /**
- * @brief Captures N-elements before the capture range lower bound, minus a delay period
+ * @brief Captures one one element before the capture range lower bound; one element after the capture range upper
+ * bound.
  *
- *        All older elements are removed.
+ *        All elements in between are also captured. All older elements are removed.
  *
  * @tparam DispatchT  data dispatch type
  * @tparam LockPolicyT  a BasicLockable (https://en.cppreference.com/w/cpp/named_req/BasicLockable) object or NoLock or
@@ -31,36 +30,32 @@ template <
   typename LockPolicyT = NoLock,
   typename ContainerT = DefaultContainer<DispatchT>,
   typename QueueMonitorT = DefaultDispatchQueueMonitor>
-class CountBefore : public Follower<CountBefore<DispatchT, LockPolicyT, ContainerT, QueueMonitorT>>
+class Ranged : public Follower<Ranged<DispatchT, LockPolicyT, ContainerT, QueueMonitorT>>
 {
 public:
-  /// Integer size type
-  using size_type = typename CaptorTraits<CountBefore>::size_type;
-
   /// Data stamp type
-  using stamp_type = typename CaptorTraits<CountBefore>::stamp_type;
+  using stamp_type = typename CaptorTraits<Ranged>::stamp_type;
+
+  /// Integer size type
+  using size_type = typename CaptorTraits<Ranged>::size_type;
 
   /// Data stamp duration type
-  using offset_type = typename CaptorTraits<CountBefore>::offset_type;
+  using offset_type = typename CaptorTraits<Ranged>::offset_type;
 
   /**
    * @brief Setup constructor
    *
-   * @param count  number of elements before to capture
    * @param delay  the delay with which to capture
    * @param container  container object with some initial state
    * @param queue_monitor  queue monitor with some initial state
-   *
-   * @throws <code>std::invalid_argument</code> if <code>count == 0</code>
    */
-  CountBefore(
-    const size_type count,
+  explicit Ranged(
     const offset_type& delay,
     const ContainerT& container = ContainerT{},
     const QueueMonitorT& queue_monitor = QueueMonitorT{});
 
 private:
-  using PolicyType = Follower<CountBefore<DispatchT, LockPolicyT, ContainerT, QueueMonitorT>>;
+  using PolicyType = Follower<Ranged<DispatchT, LockPolicyT, ContainerT, QueueMonitorT>>;
   friend PolicyType;
 
   /**
@@ -69,13 +64,14 @@ private:
    * @param[out] output  output data iterator
    * @param[in] range  data capture/sequencing range
    *
-   * @retval PRIMED  if there are N Dispatch elements with a sequencing stamp greater than or
-   *                 equal to the upper driving stamp, minus specified delay
-   * @retval ABORT  if capture is not possible
-   * @retval RETRY  otherwise
+   * @retval ABORT   If N-elements do not exist before <code>range.lower_stamp</code>
+   * @retval PRIMED  If N-elements exist before <code>range.lower_stamp</code> and
+   *                 M-elements exist after <code>range.upper_stamp</code>
+   * @retval RETRY   If N-elements exist before <code>range.lower_stamp</code> but
+   *                 M-elements do not exist after <code>range.upper_stamp</code>
    */
   template <typename OutputDispatchIteratorT>
-  inline State capture_follower_impl(OutputDispatchIteratorT output, const CaptureRange<stamp_type>& range);
+  inline State capture_follower_impl(OutputDispatchIteratorT&& output, const CaptureRange<stamp_type>& range);
 
   /**
    * @copydoc Follower::dry_capture_policy_impl
@@ -85,15 +81,27 @@ private:
   /**
    * @copydoc Follower::abort_policy_impl
    */
-  constexpr void abort_follower_impl(const stamp_type& t_abort) noexcept(true) {}
+  inline void abort_follower_impl(const stamp_type& t_abort);
 
   /**
    * @copydoc Follower::reset_policy_impl
    */
-  constexpr void reset_follower_impl() noexcept(true) {}
+  inline void reset_follower_impl() noexcept(true) {}
 
-  /// Number of elements to be caputed
-  size_type count_;
+  /**
+   * @brief Finds iterator after first in capture sequence
+   *
+   * @param range  data capture/sequencing range
+   */
+  inline auto find_after_first(const CaptureRange<stamp_type>& range) const;
+
+  /**
+   * @brief Finds iterator before last in capture sequence
+   *
+   * @param range  data capture/sequencing range
+   */
+  template <typename QueueIteratorT>
+  inline auto find_before_last(const CaptureRange<stamp_type>& range, const QueueIteratorT after_first) const;
 
   /// Capture delay
   offset_type delay_;
@@ -112,7 +120,7 @@ private:
  * @tparam QueueMonitorT queue monitor/capture preconditioning type
  */
 template <typename DispatchT, typename LockPolicyT, typename ContainerT, typename QueueMonitorT>
-struct CaptorTraits<follower::CountBefore<DispatchT, LockPolicyT, ContainerT, QueueMonitorT>>
+struct CaptorTraits<follower::Ranged<DispatchT, LockPolicyT, ContainerT, QueueMonitorT>>
     : CaptorTraitsFromDispatch<DispatchT>
 {
   /// Underlying dispatch container type
@@ -124,14 +132,14 @@ struct CaptorTraits<follower::CountBefore<DispatchT, LockPolicyT, ContainerT, Qu
   /// Thread locking policy type
   using LockPolicyType = LockPolicyT;
 
-  /// Indicates that data from this captor will NOT always be captured deterministically;
-  /// i.e. is always dependent on when data is injected, and when captrue is executed
-  static constexpr bool is_capture_deterministic = false;
+  /// Indicates that data from this captor will always be captured deterministically, so long as data
+  /// injection is monotonically sequenced
+  static constexpr bool is_capture_deterministic = true;
 };
 
 }  // namespace flow
 
 // Flow (implementation)
-#include "flow/src/follower/count_before.hpp"
+#include "flow/src/follower/ranged.hpp"
 
-#endif  // FLOW_FOLLOWER_COUNT_BEFORE_H
+#endif  // FLOW_FOLLOWER_RANGED_HPP
