--- flow/include/followers.h
@@ -1,29 +1,0 @@
-/**
- * @copyright 2020 Fetch Robotics Inc.
- * @author Brian Cairl
- *
- * @file followers.h
- */
-#ifndef FLOW_FOLLOWERS_H
-#define FLOW_FOLLOWERS_H
-
-// Flow
-#include <flow/follower/any_before.h>
-#include <flow/follower/before.h>
-#include <flow/follower/closest_before.h>
-#include <flow/follower/count_before.h>
-#include <flow/follower/latched.h>
-#include <flow/follower/matched_stamp.h>
-#include <flow/follower/ranged.h>
-
-namespace flow
-{
-/**
- * @brief Synchronization buffers with sequencing range-dependent policies
- */
-namespace follower
-{}
-
-}  // flow
-
-#endif  // FLOW_FOLLOWERS_H
