+++ flow/include/followers.hpp
@@ -1,0 +1,27 @@
+/**
+ * @copyright 2020 Fetch Robotics Inc.
+ * @author Brian Cairl
+ */
+#ifndef FLOW_FOLLOWERS_HPP
+#define FLOW_FOLLOWERS_HPP
+
+// Flow
+#include <flow/follower/any_before.hpp>
+#include <flow/follower/before.hpp>
+#include <flow/follower/closest_before.hpp>
+#include <flow/follower/count_before.hpp>
+#include <flow/follower/latched.hpp>
+#include <flow/follower/matched_stamp.hpp>
+#include <flow/follower/ranged.hpp>
+
+namespace flow
+{
+/**
+ * @brief Synchronization buffers with sequencing range-dependent policies
+ */
+namespace follower
+{}
+
+}  // flow
+
+#endif  // FLOW_FOLLOWERS_HPP
