+++ flow/test/unit/driver/next.cpp
@@ -13,8 +13,8 @@
 #include <gtest/gtest.h>
 
 // Flow
-#include <flow/captor/nolock.h>
-#include <flow/driver/next.h>
+#include <flow/captor/nolock.hpp>
+#include <flow/driver/next.hpp>
 
 using namespace flow;
 using namespace flow::driver;
