+++ flow/test/unit/follower/any_before.cpp
@@ -12,8 +12,8 @@
 #include <gtest/gtest.h>
 
 // Flow
-#include <flow/captor/nolock.h>
-#include <flow/follower/any_before.h>
+#include <flow/captor/nolock.hpp>
+#include <flow/follower/any_before.hpp>
 
 using namespace flow;
 using namespace flow::follower;
