+++ flow/test/unit/follower/ranged.cpp
@@ -14,8 +14,8 @@
 #include <gtest/gtest.h>
 
 // Flow
-#include <flow/captor/nolock.h>
-#include <flow/follower/ranged.h>
+#include <flow/captor/nolock.hpp>
+#include <flow/follower/ranged.hpp>
 
 using namespace flow;
 using namespace flow::follower;
