+++ flow/test/unit/synchronizer_st_example.cpp
@@ -16,10 +16,10 @@
 #include <gtest/gtest.h>
 
 // Flow
-#include <flow/captor/nolock.h>
-#include <flow/drivers.h>
-#include <flow/followers.h>
-#include <flow/synchronizer.h>
+#include <flow/captor/nolock.hpp>
+#include <flow/drivers.hpp>
+#include <flow/followers.hpp>
+#include <flow/synchronizer.hpp>
 
 using namespace flow;
